+++ src/main/java/com/felipehogrefe/expenses/domain/MonthExpense.java
@@ -1,6 +1,7 @@
 package com.felipehogrefe.expenses.domain;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 import javax.persistence.Entity;
 import javax.persistence.Id;
@@ -13,11 +14,11 @@ public class MonthExpense implements Serializable{
 	private Integer id;
 	
 	private String movimentation_month;
-	private double total;
+	private BigDecimal total;
 	
 	public MonthExpense() {}
 
-	public MonthExpense(Integer id, String movimentation_month, double total) {
+	public MonthExpense(Integer id, String movimentation_month, BigDecimal total) {
 		super();
 		this.id = id;
 		this.movimentation_month = movimentation_month;
@@ -65,11 +66,11 @@ public class MonthExpense implements Serializable{
 		this.movimentation_month = movimentation_month;
 	}
 
-	public double getTotal() {
+	public BigDecimal getTotal() {
 		return total;
 	}
 
-	public void setTotal(double total) {
+	public void setTotal(BigDecimal total) {
 		this.total = total;
 	}
